summarizer/app/requirements.txt
certifi
charset-normalizer
colorama
filelock
fsspec
huggingface-hub
idna
Jinja2
lxml
Markdown
MarkupSafe
mpmath
networkx
numpy
packaging
pillow
PyPDF2
python-docx
PyYAML
regex
requests
safetensors
setuptools
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
urllib3
